Adds landing page panel
Adds landing page panel which will display on page load.
Adds the toggling for displaying the page on log out and hiding the
page on login

This page was added after thinking through the design a bit more and
giving the casual visitor the opportunity to make use of some features
without logging in.