Add navigation to access Add Beverage form
Adds code to access the Add Beverage panel and the add Beverage form.

Still have a problem where I must rely on events.js for handling, but
code does work.

Once access was made available to the add Beverage form, the form was
tested and found to be working (a beverage was added)

Next step will be to link up the rest of the navigation buttons